Unify the docked and floating tag-editor panes
The tag-pane body was duplicated: ~230 near-identical lines in the docked
editor and again in each floating window, both assembling the same 40-field
FieldEditContext and applying the same seven deferred-op batches. Splitting a
view means rendering that body more than once per frame, so it has to be one
function first.

Extract Baboon::draw_tag_pane. It takes the document out of parsed_tags for
the duration and puts it back before returning; owning the document outright
is what frees the rest of the body to borrow self's other fields, which the
deferred-op plumbing needs. The docked path previously used get_mut and the
floating path already used remove/insert, so this standardizes on the latter.

The `scope` argument salts every widget id underneath, so a tag shown in two
panes will keep independent scroll/focus/collapse state while sharing one
document. Callers still pass "docked"/"floating", which combined with the tag
key are already unique today; per-pane scopes arrive with tiled splits.

No behavior change: same bars in the same order (the floating pane still has
no keyword bar), same missing-document handling, same edit application.
